+++ src/Models/HttpLoggerRequest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Ilyes512\HttpLogger\Eloquent;
+namespace Ilyes512\HttpLogger\Models;
 
 use Illuminate\Database\Eloquent\Model;
